Improve access grid to pick better outside directions
The outside direction logic is there to pick which direction to go
to when starting a line. The closest direction when starting from
a corner may be a silly one; the new system picks allows one of two
closest outside directions so that it doesn't so eagerly start
upwards when the better direction might be horizontal.